[FIX] Validation Count feature on relations (#25)
This PR fixes validation count features on relations.

This is mainly a simple fix, however, the case where update is done with "create/update/delete" is more complex, as we don't know the actual number of itemps on the final item. To fix this issue, the current item has to be fetched from the db.

Tests have been added to cover this situation.